DAHDI_NET: Report alarms as lost carrier to net_device
It recently occurre to me that for some strange reason the line/link
status of the E1 span was not reflected in the network device link
status.  This is odd and leads to strange behavior.

In general, it is assumed that a net_device only exhibits LOWER_UP
if there actually is a physical connection.  If you disconnect an
ethernet cable, LOWER_UP will be gone.  Same should happen on
DAHDI_NET devices.