Introduce ioctl() for simulating alarms under userspace control
For automatic testing scenarios, it is very useful to be able to
artificially create line alarms.  This allows testing of the whole
stack through DAHDI itself, DAHDI-interfaces in applications and
actual applications.

It introduces a new DAHDI_SIMULATE_ALARM ioctl() whcih has to be
issued on /dev/dahdi/ctl and which can be used to set or remove
any simulated alarms on the given span.