icE1usb: raise NOTOPEN alarm if 'startup' fails
This is what happens if there is insufficient isochronous USB bandwidth
to open a port: The STARTUP ioctl fails.  However, as the prior
SPANCONFIG ioctl succeeds, the line apears up with no alarms.

So let's set the NOTOPEN alarm whenever there are errors in STARTUP,
to give a clear indication to the user the port could not be opened.