icE1usb: Fix inverted logic for clock timing source
The 'struct dahdi_lineconfig.sync' parameter originates from the
'timing' field of system.conf.  dahdi_cfg passes it via the SPANCONFIG
ioctl into the kernel.

The DAHDI docs say:
> source of the master clock. If you choose 0, the port will never
> be used as a source of timing. This is appropriate when you know the
> far end should always be a slave to you.

So if '0' we should use the local clock, and if > 0, the recovered
remote clock.

Thanks to Christoph Lauter for pointing this out.

Change-Id: If35a5a52c094129c342d0f658f496b488d1826ad