dahdi-base: Allow reselection of the same clock master
__dahdi_find_master_span is called from a bunch of different places,
like when a span comes on-line, this is to potentially select a new
master.

However the way it's written now, when it's called, it will ALWAYS
tried to find a DIFFERENT master than what is selected now if there
is another available, no matter is the current one is earlier in the
list.

Change-Id: Ibe2365953c5ac7f3dcc7c10526b9eb797ea320ee
Signed-off-by: Sylvain Munaut <tnt@246tNt.com>